San Diego Competition - Autonomous changed for simplistic default values. Smart Dashboard edited.

+++ src/Commands/AutonomousCommand.h
@@ -36,9 +36,9 @@ public:
 	void CrossWithFlapper();
 	bool CheckAccelerometer();
 
-	int positionDial;
-	int autonDial;
-	int obstacle;
+	double positionDial;
+	double autonDial;
+	double obstacle;
 
 
 private:
